+++ tests/test_.py
@@ -1,8 +1,8 @@
-# -*- coding: utf-8 -*-
-import pytest
-
 import shutil
 import subprocess
+
+import pytest
+
 import tooncher
 
 
@@ -20,8 +20,3 @@ def test_start_engine():
     env = stdout.strip().split(b"\n")
     assert b"TTR_GAMESERVER=gameserver" in env
     assert b"TTR_PLAYCOOKIE=cookie" in env
-
-
-def test_api_request_invasions():
-    resp_data = tooncher.api_request(tooncher.INVASIONS_API_URL)
-    assert "invasions" in resp_data